openllama with rope scaling

+++ simplepeft/models/__init__.py
@@ -116,7 +116,9 @@ def get_model(
 
             if bnb_compatible:
                 kwargs["quantization_config"] = BitsAndBytesConfig(
-                    load_in_4bit=True, bnb_4bit_use_double_quant=True
+                    load_in_4bit=True,
+                    bnb_4bit_use_double_quant=True,
+                    bnb_4bit_compute_dtype=torch.float16,
                 )
 
             # load the pre-trained model and check if its 8-bit compatible
